requirements.txt
# Core Computer Vision & Deep Learning
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
mediapipe==0.10.8
torch==2.1.0
torchvision==0.16.0
tensorflow==2.15.0

# 3D Reconstruction & Processing
open3d==0.18.0
trimesh==4.0.5
pytorch3d==0.7.5
scipy==1.11.4
scikit-image==0.22.0

# Body Model & Human Reconstruction
smplx==0.1.28
chumpy==0.70
pillow==10.1.0

# Pose Estimation & Segmentation
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
mmpose==1.3.0
mmcv==2.1.0
mmengine==0.10.0

# Depth Estimation
timm==0.9.12

# Numerical & Scientific Computing
numpy==1.24.3
pandas==2.1.4

# Visualization & UI
matplotlib==3.8.2
plotly==5.18.0
tqdm==4.66.1

# API & Web (optional)
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2

# Utilities
pyyaml==6.0.1
requests==2.31.0
imageio==2.33.0

